Rip out deduplication and automatic logging.
