+++ package.json
@@ -24,7 +24,6 @@
     "blob-util": "^1.3",
     "chrome-webstore-upload-cli": "^1.1.1",
     "classnames": "^2.2.5",
-    "diff-match-patch": "^1.0.0",
     "docuri": "^4.2.2",
     "freeze-dry": "^0.2.0",
     "history": "^4.6.3",
