--- src/activity-logger/content_script/index.js
@@ -1,3 +1,0 @@
-// Report interactions with the page
-
-// TODO
