+++ src/content_script.js
@@ -1,2 +1,1 @@
-import 'src/activity-logger/content_script'
 import 'src/page-analysis/content_script'
